+++ lib/internal/main/repl.js
@@ -11,7 +11,7 @@ const {
   evalScript
 } = require('internal/process/execution');
 
-const { print, kStderr, kStdout } = require('internal/util/print');
+const console = require('internal/console/global');
 
 const { getOptionValue } = require('internal/options');
 
@@ -21,12 +21,14 @@ markBootstrapComplete();
 
 // --input-type flag not supported in REPL
 if (getOptionValue('--input-type')) {
-  print(kStderr, 'Cannot specify --input-type for REPL');
+  // If we can't write to stderr, we'd like to make this a noop,
+  // so use console.error.
+  console.error('Cannot specify --input-type for REPL');
   process.exit(1);
 }
 
-print(kStdout, `Welcome to Node.js ${process.version}.\n` +
-               'Type ".help" for more information.');
+console.log(`Welcome to Node.js ${process.version}.\n` +
+  'Type ".help" for more information.');
 
 const cliRepl = require('internal/repl');
 cliRepl.createInternalRepl(process.env, (err, repl) => {
