+++ lib/internal/modules/cjs/loader.js
@@ -819,9 +819,13 @@ Module.runMain = function() {
         true /* fromPromise */
       );
     });
+    // Handle any nextTicks added in the first tick of the program
+    process._tickCallback();
     return;
   }
   Module._load(process.argv[1], null, true);
+  // Handle any nextTicks added in the first tick of the program
+  process._tickCallback();
 };
 
 Module.createRequireFromPath = (filename) => {
