+++ lib/internal/process/execution.js
@@ -35,22 +35,26 @@ function tryGetCwd() {
   }
 }
 
-function evalModule(source, printResult) {
+function evalModule(source, print) {
+  const { log, error } = require('internal/console/global');
   const { decorateErrorStack } = require('internal/util');
   const asyncESM = require('internal/process/esm_loader');
-  const { kStdout, kStderr, print } = require('internal/util/print');
   asyncESM.loaderPromise.then(async (loader) => {
     const { result } = await loader.eval(source);
-    if (printResult) { print(kStdout, result); }
+    if (print) {
+      log(result);
+    }
   })
   .catch((e) => {
     decorateErrorStack(e);
-    print(kStderr, e);
+    error(e);
     process.exit(1);
   });
+  // Handle any nextTicks added in the first tick of the program.
+  process._tickCallback();
 }
 
-function evalScript(name, body, breakFirstLine, printResult) {
+function evalScript(name, body, breakFirstLine, print) {
   const CJSModule = require('internal/modules/cjs/loader');
   const { kVmBreakFirstLineSymbol } = require('internal/util');
 
@@ -75,10 +79,12 @@ function evalScript(name, body, breakFirstLine, printResult) {
         [kVmBreakFirstLineSymbol]: ${!!breakFirstLine}
       });\n`;
   const result = module._compile(script, `${name}-wrapper`);
-  if (printResult) {
-    const { kStdout, print } = require('internal/util/print');
-    print(kStdout, result);
+  if (print) {
+    const { log } = require('internal/console/global');
+    log(result);
   }
+  // Handle any nextTicks added in the first tick of the program.
+  process._tickCallback();
 }
 
 const exceptionHandlerState = { captureFn: null };
