+++ lib/internal/process/task_queues.js
@@ -49,7 +49,6 @@ function setHasTickScheduled(value) {
 
 const queue = new FixedQueue();
 
-// Should be in sync with RunNextTicksNative in node_task_queue.cc
 function runNextTicks() {
   if (!hasTickScheduled() && !hasRejectionToWarn())
     runMicrotasks();
