+++ src/node_task_queue.cc
@@ -2,7 +2,6 @@
 #include "node.h"
 #include "node_errors.h"
 #include "node_internals.h"
-#include "node_process.h"
 #include "v8.h"
 
 #include <atomic>
@@ -37,20 +36,6 @@ static void EnqueueMicrotask(const FunctionCallbackInfo<Value>& args) {
   CHECK(args[0]->IsFunction());
 
   isolate->EnqueueMicrotask(args[0].As<Function>());
-}
-
-// Should be in sync with runNextTicks in internal/process/task_queues.js
-bool RunNextTicksNative(Environment* env) {
-  TickInfo* tick_info = env->tick_info();
-  if (!tick_info->has_tick_scheduled() && !tick_info->has_rejection_to_warn())
-    env->isolate()->RunMicrotasks();
-  if (!tick_info->has_tick_scheduled() && !tick_info->has_rejection_to_warn())
-    return true;
-
-  Local<Function> callback = env->tick_callback_function();
-  CHECK(!callback.IsEmpty());
-  return !callback->Call(env->context(), env->process_object(), 0, nullptr)
-              .IsEmpty();
 }
 
 static void RunMicrotasks(const FunctionCallbackInfo<Value>& args) {
