--- test/message/async_error_nexttick_main.js
@@ -1,13 +1,0 @@
-'use strict';
-require('../common');
-const four = require('../fixtures/async-error');
-
-async function main() {
-  try {
-    await four();
-  } catch (e) {
-    console.error(e);
-  }
-}
-
-process.nextTick(main);
