--- test/message/async_error_nexttick_main.out
@@ -1,7 +1,0 @@
-Error: test
-    at one (*fixtures*async-error.js:4:9)
-    at two (*fixtures*async-error.js:17:9)
-    at processTicksAndRejections (internal/process/task_queues.js:*:*)
-    at async three (*fixtures*async-error.js:20:3)
-    at async four (*fixtures*async-error.js:24:3)
-    at async main (*message*async_error_nexttick_main.js:7:5)
