--- test/message/async_error_sync_esm.mjs
@@ -1,14 +1,0 @@
-// Flags: --experimental-modules
-/* eslint-disable node-core/required-modules */
-import '../common/index.mjs';
-import four from '../fixtures/async-error.js';
-
-async function main() {
-  try {
-    await four();
-  } catch (e) {
-    console.error(e);
-  }
-}
-
-main();
