--- test/message/async_error_sync_esm.out
@@ -1,7 +1,0 @@
-(node:*) ExperimentalWarning: The ESM module loader is experimental.
-Error: test
-    at one (*fixtures*async-error.js:4:9)
-    at two (*fixtures*async-error.js:17:9)
-    at async three (*fixtures*async-error.js:20:3)
-    at async four (*fixtures*async-error.js:24:3)
-    at async main (*message*async_error_sync_esm.mjs:8:5)
