--- test/message/async_error_sync_main.out
@@ -1,6 +1,0 @@
-Error: test
-    at one (*fixtures*async-error.js:4:9)
-    at two (*fixtures*async-error.js:17:9)
-    at async three (*fixtures*async-error.js:20:3)
-    at async four (*fixtures*async-error.js:24:3)
-    at async main (*message*async_error_sync_main.js:7:5)
