+++ test/message/unhandled_promise_trace_warnings.out
@@ -9,6 +9,9 @@
     at *
     at *
     at *
+    at *
+    at *
+    at *
 (node:*) Error: This was rejected
     at * (*test*message*unhandled_promise_trace_warnings.js:*)
     at *
@@ -18,6 +21,9 @@
     at *
     at *
 (node:*) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.
+    at *
+    at *
+    at *
     at *
     at *
     at *
